# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/availability/providers/{providerId}/weekly-schedule": {
      "get": {
        "operationId": "AvailabilityController_getWeeklySchedule",
        "summary": "Get a provider weekly schedule",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use provider or admin to manage schedules; user to view availability.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/WeeklySchedule"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "availability"
        ]
      },
      "put": {
        "operationId": "AvailabilityController_saveWeeklySchedule",
        "summary": "Replace a provider recurring weekly schedule",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use provider or admin to manage schedules; user to view availability.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertWeeklyScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/WeeklySchedule"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "availability"
        ]
      }
    },
    "/api/availability/providers/{providerId}/date-overrides": {
      "get": {
        "operationId": "AvailabilityController_getDateOverrides",
        "summary": "List a provider date overrides",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use provider or admin to manage schedules; user to view availability.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DateOverride"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "availability"
        ]
      }
    },
    "/api/availability/providers/{providerId}/date-overrides/{date}": {
      "put": {
        "operationId": "AvailabilityController_saveDateOverride",
        "summary": "Create or replace a date override",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use provider or admin to manage schedules; user to view availability.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertDateOverrideDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DateOverride"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "availability"
        ]
      },
      "delete": {
        "operationId": "AvailabilityController_deleteDateOverride",
        "summary": "Remove a date override",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use provider or admin to manage schedules; user to view availability.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DateOverride"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "availability"
        ]
      }
    },
    "/api/availability/providers/{providerId}": {
      "get": {
        "operationId": "AvailabilityController_getAvailability",
        "summary": "Generate customer-bookable availability for the configured booking window",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use provider or admin to manage schedules; user to view availability.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProviderAvailability"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "availability"
        ]
      }
    },
...
  },
  "info": {
    "title": "ServeEase API",
    "description": "In-memory NestJS backend with CRUD, DTO validation, RBAC, and Swagger.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "Role-based access header. Use \"admin\" for full access or \"user\" for read-only GET access."
      }
    },
    "schemas": {
...
      "WeeklySchedule": {
        "type": "object",
        "properties": {
          "providerId": {
            "type": "string",
            "example": "cleanpro-service"
          },
          "weeklySchedule": {
            "type": "object",
            "example": {
              "monday": [
                "09:00-12:00",
                "13:00-17:00"
              ],
              "tuesday": [
                "09:00-12:00"
              ],
              "wednesday": [],
              "thursday": [],
              "friday": [],
              "saturday": [],
              "sunday": []
            }
          }
        },
        "required": [
          "providerId",
          "weeklySchedule"
        ]
      },
      "DateOverride": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-08-04"
          },
          "fullDayOff": {
            "type": "boolean",
            "example": false
          },
          "disabledSlots": {
            "example": [
              "09:00-12:00"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "enabledSlots": {
            "example": [
              "13:00-17:00"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "date",
          "fullDayOff",
          "disabledSlots",
          "enabledSlots"
        ]
      },
      "AvailableDate": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-08-04"
          },
          "dayOfWeek": {
            "type": "string",
            "example": "Monday"
          },
          "slots": {
            "example": [
              "13:00-17:00"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "date",
          "dayOfWeek",
          "slots"
        ]
      },
      "ProviderAvailability": {
        "type": "object",
        "properties": {
          "providerId": {
            "type": "string",
            "example": "cleanpro-service"
          },
          "bookingWindowDays": {
            "type": "number",
            "example": 7
          },
          "dates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AvailableDate"
            }
          }
        },
        "required": [
          "providerId",
          "bookingWindowDays",
          "dates"
        ]
      },
      "WeeklyScheduleSlotsDto": {
        "type": "object",
        "properties": {
          "monday": {
            "example": [
              "09:00-12:00",
              "13:00-17:00"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tuesday": {
            "example": [
              "09:00-12:00"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "wednesday": {
            "example": [
              "09:00-12:00"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "thursday": {
            "example": [
              "09:00-12:00"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "friday": {
            "example": [
              "09:00-12:00"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "saturday": {
            "example": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sunday": {
            "example": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UpsertWeeklyScheduleDto": {
        "type": "object",
        "properties": {
          "weeklySchedule": {
            "example": {
              "monday": [
                "09:00-12:00",
                "13:00-17:00"
              ],
              "tuesday": [
                "09:00-12:00"
              ],
              "wednesday": [],
              "thursday": [],
              "friday": [],
              "saturday": [],
              "sunday": []
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/WeeklyScheduleSlotsDto"
              }
            ]
          }
        },
        "required": [
          "weeklySchedule"
        ]
      },
      "UpsertDateOverrideDto": {
        "type": "object",
        "properties": {
          "fullDayOff": {
            "type": "boolean",
            "example": false
          },
          "disabledSlots": {
            "example": [
              "09:00-12:00"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "enabledSlots": {
            "example": [
              "13:00-17:00"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "fullDayOff"
        ]
      },
...
    }
  }
}